Consolidate and remove duplicated UI tests
When the new UI was made the default, and only, UI to use, some 'new ui' tests still existed and needed moving or removing.
