+++ lib/nerves_hub_web/live/org/signing_keys.html.heex
@@ -1,5 +1,5 @@
 <%= if @live_action == :index do %>
-  <div class="h-[56px] flex justify-end bg-base-900 border-b border-base-700"></div>
+  <div class="h-14 flex justify-end bg-base-900 border-b border-base-700"></div>
 
   <div class="h-0 flex-1 overflow-y-auto">
     <div class="flex items-center h-[90px] gap-4 px-6 py-7 border-b border-[#3F3F46] text-sm font-medium">
@@ -47,7 +47,7 @@
 <% end %>
 
 <%= if @live_action == :new do %>
-  <div class="h-[56px] flex justify-end bg-base-900 border-b border-base-700"></div>
+  <div class="h-14 flex justify-end bg-base-900 border-b border-base-700"></div>
 
   <div class="h-0 flex-1 overflow-y-auto">
     <div class="flex items-center h-[90px] gap-4 px-6 py-7 border-b border-[#3F3F46] text-sm font-medium">
