+++ test/nerves_hub_web/live/deployment_groups/show/releases_tab_test.exs
@@ -1,4 +1,4 @@
-defmodule NervesHubWeb.Live.NewUI.DeploymentGroups.Show.ReleasesTabTest do
+defmodule NervesHubWeb.Live.DeploymentGroups.Show.ReleasesTabTest do
   use NervesHubWeb.ConnCase.Browser, async: false
   use Mimic
