+++ test/nerves_hub_web/live/devices/show/local_shell_tab_test.exs
@@ -1,4 +1,4 @@
-defmodule NervesHubWeb.Live.NewUI.Devices.LocalShellTabTest do
+defmodule NervesHubWeb.Live.Devices.Show.LocalShellTabTest do
   use NervesHubWeb.ConnCase.Browser, async: false
 
   alias NervesHub.Devices
