+++ test/nerves_hub_web/live/support_scripts_test.exs
@@ -4,8 +4,9 @@ defmodule NervesHubWeb.Live.SupportScriptsTest do
   alias NervesHub.Fixtures
   alias NervesHub.Scripts
 
-  setup %{user: user, org: org} do
+  setup %{user: user, org: org} = context do
     [product: Fixtures.product_fixture(user, org, %{name: "Amazing"})]
+    context
   end
 
   describe "list" do
@@ -16,7 +17,7 @@ defmodule NervesHubWeb.Live.SupportScriptsTest do
     } do
       conn
       |> visit("/org/#{org.name}/#{product.name}/scripts")
-      |> assert_has("span", text: "#{product.name} doesn’t have any Support Scripts.")
+      |> assert_has("span", text: "#{product.name} doesn’t have any Support Scripts")
     end
 
     test "shows all support scripts for a product", %{
@@ -30,6 +31,57 @@ defmodule NervesHubWeb.Live.SupportScriptsTest do
       conn
       |> visit("/org/#{org.name}/#{product.name}/scripts")
       |> assert_has("td", text: "MOTD")
+    end
+  end
+
+  describe "pagination" do
+    test "no pagination when less than 25 support scripts", %{
+      conn: conn,
+      org: org,
+      product: product,
+      user: user
+    } do
+      {:ok, _script} = Scripts.create(product, user, %{name: "MOTD", text: "NervesMOTD.print()"})
+
+      conn
+      |> visit("/org/#{org.name}/#{product.name}/scripts")
+      |> refute_has("button", text: "25", timeout: 1000)
+    end
+
+    test "paginate when more than 25 support scripts", %{
+      conn: conn,
+      org: org,
+      product: product,
+      user: user
+    } do
+      for i <- 1..26 do
+        {:ok, _script} =
+          Scripts.create(product, user, %{name: "MOTD #{i}", text: "NervesMOTD.print()"})
+      end
+
+      scripts = Scripts.all_by_product(product)
+      [first_script | _] = scripts |> Enum.sort_by(& &1.name)
+
+      conn
+      |> visit("/org/#{org.name}/#{product.name}/scripts")
+      |> assert_has("button", text: "25", timeout: 1000)
+      |> assert_has("button", text: "50", timeout: 1000)
+      |> assert_has("button", text: "2", timeout: 1000)
+      |> click_button("button[phx-click='paginate'][phx-value-page='2']", "2")
+      |> refute_has("a", text: first_script.name, exact: true, timeout: 1000)
+    end
+  end
+
+  describe "delete" do
+    test "removes support script", %{conn: conn, org: org, product: product, user: user} do
+      {:ok, script} = Scripts.create(product, user, %{name: "MOTD", text: "NervesMOTD.print()"})
+
+      conn
+      |> visit("/org/#{org.name}/#{product.name}/scripts/#{script.id}/edit")
+      |> assert_has("input", value: script.name)
+      |> click_button("Delete Script")
+      |> assert_path("/org/#{org.name}/#{product.name}/scripts")
+      |> assert_has("span", text: "#{product.name} doesn’t have any Support Scripts")
     end
   end
 
@@ -49,6 +101,19 @@ defmodule NervesHubWeb.Live.SupportScriptsTest do
       |> assert_path("/org/#{org.name}/#{product.name}/scripts")
       |> assert_has("td", text: "MOTD")
     end
+
+    test "add script with tags", %{conn: conn, org: org, product: product} do
+      conn
+      |> visit("/org/#{org.name}/#{product.name}/scripts/new")
+      |> fill_in("Name", with: "MOTD")
+      |> fill_in("Script code", with: "NervesMOTD.print()")
+      |> fill_in("Tags", with: "hello,world")
+      |> click_button("Save changes")
+      |> assert_path("/org/#{org.name}/#{product.name}/scripts")
+      |> assert_has("td", text: "MOTD")
+      |> assert_has("span", text: "hello")
+      |> assert_has("span", text: "world")
+    end
   end
 
   describe "edit" do
@@ -64,9 +129,12 @@ defmodule NervesHubWeb.Live.SupportScriptsTest do
       |> assert_has("p", text: "can't be blank", count: 1)
       |> fill_in("Name", with: "MOTD")
       |> fill_in("Script code", with: "dbg(NervesMOTD.print())")
+      |> fill_in("Tags", with: "hello,world")
       |> click_button("Save changes")
       |> assert_path("/org/#{org.name}/#{product.name}/scripts")
       |> assert_has("td", text: "MOTD")
+      |> assert_has("span", text: "hello")
+      |> assert_has("span", text: "world")
 
       assert %{text: "dbg(NervesMOTD.print())"} = Scripts.get!(script.id)
     end
